requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Task queue
celery==5.3.6
redis==5.0.4

# HTTP
aiohttp==3.9.5
httpx==0.27.0

# Domain analysis
python-Levenshtein==0.25.1
tldextract==5.1.2
dnspython==2.6.1
python-whois==0.9.4

# Visual analysis
playwright==1.43.0
imagehash==4.3.1
Pillow==10.3.0
colorthief==0.2.1

# Content analysis
beautifulsoup4==4.12.3
lxml==5.2.1

# Certstream
websockets==12.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.2.3
